requirements.txt
mcp>=0.9.0
psycopg2-binary>=2.9.0
jira>=3.5.0
python-dotenv>=1.0.0
nltk>=3.8.0
scikit-learn>=1.3.0
python-Levenshtein>=0.21.0
